+++ aidl-module/src/main/java/th/co/digio/aidl_module/BinderAImpl.kt
@@ -1,4 +1,4 @@
-package th.co.digio.service
+package th.co.digio.aidl_module
 
 import android.os.RemoteException
 import android.util.Log
