+++ service/src/main/java/th/co/digio/service/BinderPoolService.kt
@@ -4,6 +4,8 @@ import android.app.Service
 import android.content.Intent
 import android.os.IBinder
 import android.os.RemoteException
+import th.co.digio.aidl_module.BinderAImpl
+import th.co.digio.aidl_module.BinderBImpl
 import th.co.digio.aidl_module.IBinderPool
 
 class BinderPoolService : Service() {
